consolidate inputfile testing
